+++ src/main/java/com/codessquad/qna/web/service/UserService.java
@@ -1,0 +1,31 @@
+package com.codessquad.qna.web.service;
+
+import com.codessquad.qna.web.domain.User;
+import com.codessquad.qna.web.repository.UserRepository;
+import org.springframework.stereotype.Service;
+
+import java.util.NoSuchElementException;
+
+@Service
+public class UserService {
+
+    private UserRepository userRepository;
+
+    UserService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    public User findUserById(Long id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("해당 id의 사용자가 존재하지 않습니다."));
+        return user;
+    }
+
+    public void save(User user) {
+        userRepository.save(user);
+    }
+
+    public Iterable<User> findAll() {
+        return userRepository.findAll();
+    }
+}
